swift: extra consistency check after Move
We face a consistency issue with our data store, time to time we end up with
empty data file. This problem is similar to #2188.

The upload of a blob is done by the following steps:

a) upload in a temporary location
b) validation of the blob[1]
c) link of the blob
d) Move
e) Cache is updated

Because of e, the corrupted blob won't be change by another image push.

a, b, and c look good in our context. So we assume *something* happens time to
time during the Move() call. This patch does an extra Stat() call after
the move, to ensure the size is still the same.

[1]: https://github.com/docker/distribution/blob/16128bbac47f75050e82f7e91b04df33775e0c23/registry/storage/blobwriter.go#L69
[2]: https://github.com/docker/distribution/blob/16128bbac47f75050e82f7e91b04df33775e0c23/registry/storage/driver/swift/swift.go#L501-L517